Coolant thoughts
Coolant hasn't been working well recently. Wondering whether it's a combination of the tank being a little low, and the return to the tank being clogged enough with swarf that the return to the tank is slowed? Or maybe the coolant has become too viscous by evaporation. Will do a health check with current coolant as a baseline, and then maybe replace with a "fresh batch" of coolant and run again.